# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Silver Stream Social Media API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}/follow:
    post:
      operationId: followUser
      tags: [follows]
      summary: Follow a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
    delete:
      operationId: unfollowUser
      tags: [follows]
      summary: Unfollow a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
...
  /posts/{postId}/comments:
    get:
      operationId: getComments
      tags: [comments]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentPage"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ── STORIES ──────────────────────────────────────────
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for current user
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPage"
  /notifications/read:
    post:
      operationId: markNotificationsRead
      tags: [notifications]
      summary: Mark notifications as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadInput"
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
  /conversations:
    get:
      operationId: getConversations
      tags: [chat]
      summary: Get all conversations for current user
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [chat]
      summary: Start a conversation with a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "200":
          description: Conversation (existing or new)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{conversationId}/messages:
    get:
      operationId: getMessages
      tags: [chat]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePage"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ── SEARCH ──────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        isFollowing:
          type: boolean
        isMe:
          type: boolean
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
      required: [id, clerkId, username, followersCount, followingCount, postsCount, isFollowing, isMe, onboardingComplete, createdAt]
...
    FollowResult:
      type: object
      properties:
        following:
          type: boolean
        followersCount:
          type: integer
      required: [following, followersCount]
...
    Comment:
      type: object
      properties:
        id:
          type: string
        postId:
          type: string
        authorId:
          type: string
        author:
          $ref: "#/components/schemas/UserProfile"
        content:
          type: string
        createdAt:
          type: string
      required: [id, postId, authorId, author, content, createdAt]
    CommentInput:
      type: object
      properties:
        content:
          type: string
      required: [content]
    CommentPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        nextCursor:
          type: ["string", "null"]
      required: [items]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: ["like", "dislike", "comment", "follow", "mention"]
        actorId:
          type: string
        actor:
          $ref: "#/components/schemas/UserProfile"
        postId:
          type: ["string", "null"]
        commentId:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, type, actorId, actor, isRead, createdAt]
    NotificationPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        nextCursor:
          type: ["string", "null"]
      required: [items]
    MarkReadInput:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
        all:
          type: boolean
...
    Conversation:
      type: object
      properties:
        id:
          type: string
        participant:
          $ref: "#/components/schemas/UserProfile"
        lastMessage:
          type: ["object", "null"]
          properties:
            content:
              type: string
            createdAt:
              type: string
            isMe:
              type: boolean
        unreadCount:
          type: integer
        updatedAt:
          type: string
      required: [id, participant, unreadCount, updatedAt]
    ConversationInput:
      type: object
      properties:
        participantId:
          type: string
      required: [participantId]
    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        sender:
          $ref: "#/components/schemas/UserProfile"
        content:
          type: string
        isMe:
          type: boolean
        createdAt:
          type: string
      required: [id, conversationId, senderId, sender, content, isMe, createdAt]
    MessageInput:
      type: object
      properties:
        content:
          type: string
      required: [content]
    MessagePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        nextCursor:
          type: ["string", "null"]
      required: [items]
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...